==> Code/00_File_Template.R <==
# Set-up =======================================================================

#' have all the source calls etc in the top of the file 

# Section 1 ====================================================================

#' In each section, all code should be in as few chunks as possible and as much 
#' in base code as possible (fewer dependency issues)

# Last Section =================================================================

#' Ensure that any writing out of objects all happens at the end and that we 
#' do any small tests here
